Add files via upload
Added mixed coordination for n+1 and n-1.  With this added option to have the mc minimisation to change node coordination number.  As n-1 sites in the node coordination = 3 can be either bent or linear, added an option to specify the fraction of n-1 sites that are linear, but this is not yet functional.  Most additions (very poorly) commented.  Code now writes out the mixed coordination fractions from the input and the number of each type of site.